\name{TitanCNA trained dataset}
\docType{data}

\alias{TitanCNA-dataset}
\alias{EMresults}
\alias{convergeParams}
\alias{data}
\title{TITAN EM trained results for an example dataset}
\usage{
  data(EMresults)
}

\description{
  Data for chromosome 2 for a triple-negative breast cancer dataset and the expectation-maximization (EM) trained results. Only 20,000 datapoints are included and the data has been scrambled to anonymous patient SNPs.
  \describe{
    \item{data}{Processed input data that is first generated by
      \code{\link{loadAlleleCountsFromFile}}, and includes log ratios that 
      have been GC content and mappability corrected using 
      \code{\link{correctReadDepth}}}.
    \item{convergeParams}{EM results that is generated by 
      \code{\link{runEMclonalCN}}}
  }
}


\format{
  \sQuote{data} is a list.
  \sQuote{convergeParams} is a list.
}

\keyword{datasets}

