\name{TitanCNA trained dataset}
\docType{data}

\alias{TitanCNA-dataset}
\alias{EMresults}
\alias{convergeParams}
\alias{data}
\title{TITAN EM trained results for an example dataset}
\usage{
  data(EMresults)
}

\description{
  Data for chromosome 2 for a triple-negative breast cancer dataset and the expectation-maximization (EM) trained results. Only 20,000 datapoints are included and the data has been scrambled to anonymous patient SNPs.
  \describe{
    \item{data}{Processed input data that is first generated by
      \code{\link{loadAlleleCounts}}, and includes log ratios that 
      have been GC content and mappability corrected using 
      \code{\link{correctReadDepth}}}.
    \item{convergeParams}{EM results that is generated by 
      \code{\link{runEMclonalCN}}}
  }
}


\format{
  \sQuote{data} is a list.
  \sQuote{convergeParams} is a list.
}

\references{
Ha, G., Roth, A., Khattra, J., Ho, J., Yap, D., Prentice, L. M., Melnyk, N., McPherson, A., Bashashati, A., Laks, E., Biele, J., Ding, J., Le, A., Rosner, J., Shumansky, K., Marra, M. A., Huntsman, D. G., McAlpine, J. N., Aparicio, S. A. J. R., and Shah, S. P. (2014). TITAN: Inference of copy number architectures in clonal cell populations from tumour whole genome sequence data. Genome Research, Published in advance July 24 2014. (PMID: 25060187)
}

\keyword{datasets}

